Improve ux for rooms

See merge request xivdp/ef-nl-frontend!6

+++ src/views/Room/components/addEditButtons.js
@@ -5,6 +5,9 @@ import { useSelector, useDispatch } from "react-redux"
 import {setAddEditPopup, setAddEditPopupData} from '@src/redux/actions/layout'
 import { deleteRoom } from "@src/redux/actions/rooms"
 
+import Swal from 'sweetalert2'
+import withReactContent from 'sweetalert2-react-content'
+
 const AddEditBtn = ({isAdd, data}) => {
     const dispatch = useDispatch()
     const {addEditPopup} = useSelector(state => state.layout)
@@ -17,10 +20,35 @@ const AddEditBtn = ({isAdd, data}) => {
       }
     }
 
-    const handleClick = () => {
-      const {_id} = data
-      dispatch(deleteRoom(_id))
-    }
+    const handleDelete = async () => {
+        const MySwal = withReactContent(Swal)
+        const result = await MySwal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonText: 'Yes, delete it!',
+            customClass: {
+                confirmButton: 'btn btn-primary',
+                cancelButton: 'btn btn-outline-danger ml-1'
+            },
+            buttonsStyling: false
+        })
+        if (result.value) {
+            const { _id } = data
+            const deleteObj = await dispatch(deleteRoom(_id))
+            if (deleteObj?.success) {
+              MySwal.fire({
+                icon: "success",
+                title: "Room Deleted!",
+                text: "Room has been deleted.",
+                customClass: {
+                  confirmButton: "btn btn-success"
+                }
+              })
+            }
+        }
+      }
 
     return isAdd ? (
       <Col className="w-100 text-right">
@@ -32,13 +60,13 @@ const AddEditBtn = ({isAdd, data}) => {
     ) : (
       <div className="demo-inline-spacing">
         <Button.Ripple
-          className="btn-icon"
+          className="btn-icon m-0"
           color="flat-success"
           onClick={(e) => handleOpen(true)}
         >
           <FiEdit size={16} />
         </Button.Ripple>
-        <Button.Ripple className="btn-icon" onClick={(e) => handleClick()} color="flat-danger">
+        <Button.Ripple className="btn-icon m-0" onClick={(e) => handleDelete()} color="flat-danger">
           <FiTrash2 size={16} />
         </Button.Ripple>
       </div>
